Creación de Hook y utilización de React.lazy y Suspense
