+++ src/components/TrendingSearches/index.jsx
@@ -1,50 +1,20 @@
-import { useState, useEffect, useRef } from "react";
-import getTrendingTerms from '@/services/getTrendingTermsService'
-import Category from "../Category";
+import React, {Suspense} from "react";
+import useNearScreen from "@/hooks/useNearScreen";
+import Spinner from "../Spinner/index";
 
-function TrendingSearches () {
-  const [trends, setTrends] = useState([]);
 
-  useEffect(function () {
-    getTrendingTerms().then(setTrends)
-  }, [])
+const TrendingSearches = React.lazy(
+  () => import('./TrendingSearches')
+)
 
-  return <Category name='Tendencias' options={trends}/>
-
-}
 
 export default function LazyTrending () {
-  const [show, setShow] = useState(false);
-  const elementRef = useRef();
+  const {isNearScreen, fromRef} = useNearScreen({ distance: '100px' });
 
-
-  useEffect(function () {
-    let observer;
-    const onChange = (entries, observer) => {
-      const el = entries[0];
-      if(el.isIntersecting){
-        setShow(true);
-        observer.disconnect();
-      }
-    }
-
-    Promise.resolve(
-      typeof IntersectionObserver !== 'undefined' ? IntersectionObserver : import('react-intersection-observer')
-    ).then(() => {
-
-      observer = new IntersectionObserver(onChange, {
-        rootMargin: '100px'
-      })
-  
-      observer.observe(elementRef.current);
-  
-      return () => observer.disconnect();
-    })
-
-    })
-
-  return <div ref={elementRef}>
-    {show ? <TrendingSearches/> : null}
+  return <div ref={fromRef}>
+    <Suspense fallback={<Spinner/>} >
+    {isNearScreen ? <TrendingSearches/> : <Spinner/>}
+    </Suspense>
   </div>
 
 }
